// apps/web/src/lib/api/generated/models/automationDefinitionStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export type AutomationDefinitionStatusInput =
	(typeof AutomationDefinitionStatusInput)[keyof typeof AutomationDefinitionStatusInput];

export const AutomationDefinitionStatusInput = {
	active: 'active',
	paused: 'paused'
} as const;