// apps/web/src/lib/api/generated/models/automationTriggerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export type AutomationTriggerInput =
	(typeof AutomationTriggerInput)[keyof typeof AutomationTriggerInput];

export const AutomationTriggerInput = {
	report_added: 'report_added',
	acquisition_completed: 'acquisition_completed',
	full_text_attached: 'full_text_attached',
	report_included: 'report_included',
	study_created: 'study_created',
	appraisal_completed: 'appraisal_completed',
	manual: 'manual'
} as const;