// apps/web/src/lib/api/generated/models/automationDefinitionDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */
import type { AutomationStepDto } from './automationStepDto.ts';

export interface AutomationDefinitionDto {
	created_at: string;
	id: string;
	name: string;
	project_id: string;
	recipe: string;
	status: string;
	steps: AutomationStepDto[];
	trigger: string;
	updated_at: string;
	version: number;
}